[add] 差分更新の実装
[add] integratorのステータス管理クラス `JobcanDIStatus` を追加

- 進捗Enum の `.name` から、具体的な `DetailedProgressStatus` を取得する `get_detailed_progress_status_from_str` を追加
- integrator のステータス管理クラス `JobcanDIStatus` および関連クラスを追加
- 上記クラスのテストコードを追加
- TODOを更新

[update] 進捗等のステータスを app_status で管理するように変更

- 実行時は現在のステータスを表示するため、 `JobcanDIStatus` の `.previous_status` を `app_status` に変更
- 簡単のため `AppProgress` クラスに `.get` と `.set` を追加
- コンフィグファイルの読み込み先に影響するため、 `JobcanDIConfig` で `app_status` を先に読み込むように変更
- `JobcanDataIntegrator` で進捗状況を .config.app_status` を介して保持するように変更

[update] キャンセル時に sys.exit を実行しないように変更

- キャンセル時に `sys.exit` を実行しないようにした
- `.is_canceled` が `True` の場合、以降の処理を進めないようにした

[add] 申請書(概要)取得時に前回の取得日時でフィルタリング

- 申請書(概要)取得時に、前回の最終取得時刻をapplied_after条件として追加
- app_status に各 form_id ごとの最終取得日時を追加
- テストコードを更新

[add] 前回の進捗状況に応じた、実行内容のスキップ機能を実装（大枠）

- `JobcanDIStatus.load` での読み込み時、前回終了が `TerminatingStatus.COMPLETED` であれば初期化 (`ProgressStatus.INITIALIZING` & `None`) する
- 前回終了時の進捗 (正確には現在の `._current_progress`) に応じて、処理をスキップするように変更
- `JobcanDataIntegrator` のプロパティを `@property` 化

[update] 一時ファイルにデータがないときのみフォルダを削除するように変更

- form_outline の一時保存用データクラス  `TempFormOutline` を追加
- すべての一時ファイルがデータを持たない（idが空など）場合にのみ一時ファイルディレクトリを削除するように変更
- テストコードを追加

[add] form_outline の読み込みの際に、一時ファイルに読み込み済みの form_id でのスキップ機能を追加

- `AddProgress`, `FetchFailureRecord` のメンバ変数を隠蔽
- `AddProgress` に request_id などを格納するための `._specifics` メンバを追加
- form_outline の読み込みの際に、一時ファイルに読み込み済みの form_id はスキップするように変更

[update] form_outline の取得時に、前回失敗したかを考慮するように変更

- `AppProgress`, `FetchFailureRecord` の各要素を set として保持するように変更
- `AppProgress` に、指定された処理が処理済みかを判定する `.is_future_process` を追加
- form_id と request_id 等との偶発的な一致を除外するため、上記関数により前回取得済みかを判定するように変更
- `JobcanDataIntegrator._fetch_basic_data` で一つでも取得ミスがあれば取得失敗とするように変更
- 申請書(概要)取得失敗時に app_status に失敗した form_id を記録するように変更
- 前回取得に失敗した form_id について、かならず申請書(概要)を取得するように更新

[update] 申請書(概要)取得時にキャンセルされた申請書も対象とした

- `JobcanDataIntegrator.fetch_form_outline_data` がキャンセルされた申請書も取得できるように変更

[add] TempFileIO が一時ファイルを使用しないオプションを追加

- `TempFileIO` 等を `TempDataIO` に変更
- `TempDataIO` にデータをメモリにのみ格納するモードを追加

[add] form_detail の差分更新を追加

- `requests` テーブルから request_id を取得する関数 `retrieve_ids` を追加
- form_detail の差分更新を追加
- ドキュメントを更新

+++ jobcan_di/integrator/_tf_io.py
@@ -2,19 +2,101 @@
 This module provides temporary file I/O classes.
 
 Classes:
-    TempFileIO: Temporary file I/O class
-    JobcanTempFileIO: Jobcan temporary file I/O class
+    TempDataIO: Temporary data I/O class
+    JobcanTempDataIO: Jobcan temporary data I/O class
+
+Notes:
+    If `prioritize_memory` is True, the data is stored in memory and not saved to files.
 """
 import json
-from os import path, makedirs
+from os import path, makedirs, remove
 import shutil
+from typing import Dict, Union, Iterable, Optional
 
 
 
-class TempFileIO:
+class TempFormOutline:
+    """Data class for form outline (temporary)"""
+    def __init__(self,
+                 success: bool = False,
+                 ids: Union[list[str], set[str], None] = None,
+                 last_access: str = ""):
+        """Constructor
+
+        Parameters
+        ----------
+        success : bool
+            Success flag
+        ids : list[str] | set[str] | None
+            list or set of request_id
+        last_access : str
+            Last access datetime
+        """
+        self.success: bool = success
+        """Success flag"""
+        self.ids: set[str] = set(ids) if ids else set()
+        """set of request_id"""
+        self.last_access: str = last_access
+        """Last access datetime"""
+
+    def add_ids(self, ids: Iterable[str]):
+        """Add request_id to the set
+
+        Parameters
+        ----------
+        ids : Iterable[str]
+            list or set of request_id
+        """
+        self.ids.update(ids)
+
+    def remove_id(self, id_: str):
+        """Remove request_id from the set
+
+        Parameters
+        ----------
+        id_ : str
+            request_id
+        """
+        self.ids.discard(id_)
+
+    def asdict(self, json_format: bool = False) -> dict:
+        """Convert to dictionary
+
+        Parameters
+        ----------
+        json_format : bool
+            If True, convert to dictionary for JSON format
+
+        Returns
+        -------
+        dict
+            Converted dictionary
+        """
+        if json_format:
+            return {"success": self.success,
+                    "ids": list(self.ids),
+                    "last_access": self.last_access}
+        return {"success": self.success,
+                "ids": self.ids,
+                "last_access": self.last_access}
+
+    @property
+    def is_empty(self) -> bool:
+        """Check if the set of request_id is empty
+
+        Returns
+        -------
+        bool
+            True if the set is empty, otherwise False
+        """
+        return not bool(self.ids)
+
+
+
+class TempDataIO:
     """Temporary file I/O class
     """
-    def __init__(self, app_dir: str):
+    def __init__(self, app_dir: str, *, prioritize_memory: bool = False):
         """Constructor
 
         Parameters
@@ -22,10 +104,18 @@ class TempFileIO:
         app_dir : str
             Application directory path.
             All temporary files are stored under this directory.
+        prioritize_memory : bool
+            If True, prioritize memory over files (i.e. do not save to files).
         """
         # TODO: Rename files when temporary directory with different processing contents exist.
         self._temp_dir = path.join(app_dir, "temp")
         """Temporary directory path"""
+
+        self._prioritize_memory = prioritize_memory
+        """Prioritize memory over files"""
+        self._raw_data: Dict[str, Optional[dict]] = {}
+        """Raw data. The key is the file name.
+        If prioritize_memory is True or the temporary file is not available, use this data"""
 
         self._init_directories()
 
@@ -34,14 +124,73 @@ class TempFileIO:
         if not path.exists(self._temp_dir):
             makedirs(self._temp_dir)
 
+    def _save_files(self, data: dict, file_name: str):
+        """Save data to a file
+
+        Parameters
+        ----------
+        data : dict
+            Data to save. The data should be JSON serializable.
+        file_name : str
+            File name
+        """
+        if self._prioritize_memory:
+            # If prioritize_memory is True, save the data to memory
+            self._raw_data[file_name] = data
+            return
+
+        try:
+            with open(path.join(self._temp_dir, file_name), "w", encoding="utf-8") as f:
+                json.dump(data, f, indent=None, ensure_ascii=False)
+            self._raw_data[file_name] = None
+        except OSError:
+            self._raw_data[file_name] = data
+
+    def _load_files(self, file_name: str) -> dict:
+        """Load data from a file
+
+        Parameters
+        ----------
+        file_name : str
+            File name
+
+        Returns
+        -------
+        dict
+            Loaded data. {} if the file does not exist
+        """
+        if self._prioritize_memory:
+            # If prioritize_memory is True, return the data in memory
+            return self._raw_data.get(file_name, dict())
+
+        try:
+            with open(path.join(self._temp_dir, file_name), "r", encoding="utf-8") as f:
+                return json.load(f)
+        except FileNotFoundError:
+            return dict()
+        except OSError:
+            return self._raw_data.get(file_name, dict())
+
+    def _remove_files(self) -> bool:
+        """Remove temporary files if they are in the initial state
+
+        Returns
+        -------
+        bool
+            True if the files are removed, otherwise False
+        """
+        return True
+
     def cleanup(self):
         """Clear temporary directory"""
-        shutil.rmtree(self._temp_dir, ignore_errors=True)
+        if self._remove_files() or self._prioritize_memory:
+            # If the files are removed or not used, remove the directory
+            shutil.rmtree(self._temp_dir, ignore_errors=True)
 
-class JobcanTempFileIO(TempFileIO):
+class JobcanTempDataIO(TempDataIO):
     """Jobcan temporary file I/O class
     """
-    def __init__(self, app_dir: str):
+    def __init__(self, app_dir: str, *, prioritize_memory: bool = False):
         """Constructor
 
         Parameters
@@ -49,20 +198,58 @@ class JobcanTempFileIO(TempFileIO):
         app_dir : str
             Application directory path.
             All temporary files are stored under this directory.
+        prioritize_memory : bool
+            If True, prioritize memory over files (i.e. do not save to files).
         """
-        super().__init__(app_dir=app_dir)
-        self._form_outline_tmp_file = path.join(self._temp_dir,
-                                                "form_outline_temp.json")
-        """temporary file path for form outline"""
+        super().__init__(app_dir=app_dir, prioritize_memory=prioritize_memory)
+        self._form_outline_tmp_file = "form_outline_temp.json"
+        """temporary file name for form outline"""
 
-    def save_form_outline(self, data: dict):
+    def save_form_outline(self, data: Dict[int, TempFormOutline]):
         """Save form outline to temporary file
-        """
-        with open(self._form_outline_tmp_file, "w", encoding="utf-8") as f:
-            json.dump(data, f, indent=None, ensure_ascii=False)
 
-    def load_form_outline(self) -> dict:
-        """Load form outline from temporary file
+        Parameters
+        ----------
+        data : dict
+            Form outline data.
+            Key: form_id, Value: TempFormOutline
         """
-        with open(self._form_outline_tmp_file, "r", encoding="utf-8") as f:
-            return json.load(f)
+        j_data = {k: v.asdict(json_format=True) for k, v in data.items()}
+        self._save_files(j_data, self._form_outline_tmp_file)
+
+
+    def load_form_outline(self) -> Dict[int, TempFormOutline]:
+        """Load form outline from temporary file
+
+        Returns
+        -------
+        dict
+            Form outline data.
+            Key: form_id, Value: TempFormOutline
+        """
+        j_data = self._load_files(self._form_outline_tmp_file)
+
+        return {int(k): TempFormOutline(**v) for k, v in j_data.items()}
+
+    def _remove_files(self):
+        """Remove temporary files if they are in the initial state
+
+        Returns
+        -------
+        bool
+            True if the files are removed, otherwise False
+        """
+        s = super()._remove_files()
+
+        # form outline
+        if (fo := self.load_form_outline()):
+            # If the form outline is empty, remove the file
+            if all([f.is_empty for f in fo.values()]) and not self._prioritize_memory:
+                try:
+                    remove(self._form_outline_tmp_file)
+                except FileNotFoundError:
+                    pass
+            else:
+                s *= False
+
+        return s
